* rtl.h (SUBREG_PROMOTED_UNSIGNED_P): Add cast to int.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@148712 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/rtl.h
@@ -1120,7 +1120,7 @@ do {									\
 } while (0)
 #define SUBREG_PROMOTED_UNSIGNED_P(RTX)	\
   ((RTL_FLAG_CHECK1("SUBREG_PROMOTED_UNSIGNED_P", (RTX), SUBREG)->volatil) \
-     ? -1 : (RTX)->unchanging)
+   ? -1 : (int) (RTX)->unchanging)
 
 /* Access various components of an ASM_OPERANDS rtx.  */
